sidebar - active component styling and dropdown effect
