+++ src/extensions/workspace/ui/side-bar/component-tree/component-nodes/component-nodes.module.scss
@@ -15,29 +15,21 @@
 	> .arrow {
 		position: absolute; //only used to not take space
 		line-height: inherit;
-		transform: translate(-16px, 0);
+		transform: translate(-20px, 0px);
 		transition: all 300ms;
 
 		[data-collapsed="true"] & {
-			transform: translate(-16px, 2px) rotate(-0.25turn);
+			transform: translate(-20px, 1px) rotate(-0.25turn);
 		}
 	}
 }
 
-.component {
-	height: 40px;
-	line-height: 40px;
-	display: block;
-	
-	//TODO - move to a link component
-	text-decoration: none;
-	color: inherit;
-	
-	box-sizing: border-box;
-	border-left: 0px solid transparent;
-	
-	&.active {
-		background: $w20;
-		border-left: 8px $p40 solid;
+
+.componentTree {
+	max-height: 0;
+	overflow: hidden;
+	transition: max-height	400ms ease-in-out;
+	&.open {
+		max-height: 100vh;
 	}
 }
