+++ src/extensions/workspace/ui/side-bar/component-tree/component-nodes/component-nodes.tsx
@@ -1,14 +1,11 @@
-import React, { Component, useState } from 'react';
+import React, { useState } from 'react';
 import classNames from 'classnames';
-import { NavLink } from 'react-router-dom';
 import { Icon } from '@bit/bit.evangelist.elements.icon';
 
 import { TreeNodeProps, TreeLayer } from '../recursive-tree';
-import { ComponentTreeContext } from '../component-tree-context';
 import { indentStyle, indentClass } from '../indent';
 import { getName } from '../utils/get-name';
 import { clickable } from '../../../../../../to-eject/css-components/clickable';
-import { hoverable } from '../../../../../../to-eject/css-components/hoverable';
 
 import styles from './component-nodes.module.scss';
 
@@ -35,41 +32,9 @@ export function NamespaceView({ node, depth }: TreeNodeProps) {
         </div>
       )}
 
-      {!collapsed && (
-        <div style={indentStyle(depth + 1)}>
-          {node.children && <TreeLayer childNodes={node.children} depth={depth} />}
-        </div>
-      )}
+      <div style={indentStyle(depth + 1)} className={classNames(styles.componentTree, { [styles.open]: !collapsed })}>
+        {node.children && <TreeLayer childNodes={node.children} depth={depth} />}
+      </div>
     </div>
   );
-}
-export class ComponentView extends Component<TreeNodeProps> {
-  static contextType = ComponentTreeContext;
-  // context!: IComponentTreeContext;
-
-  // handleClick = () => {
-  //   const { node } = this.props;
-  //   this.context.onSelect(node.id);
-  // };
-
-  // private get isSelected() {
-  //   const { node } = this.props;
-  //   const { selected } = this.context;
-
-  //   return node.id === selected;
-  // }
-
-  render() {
-    const { node } = this.props;
-
-    return (
-      <NavLink
-        to={`/${node.id}`}
-        className={classNames(indentClass, clickable, hoverable, styles.component)}
-        activeClassName={styles.active}
-      >
-        {getName(node.id)}
-      </NavLink>
-    );
-  }
 }
