+++ src/extensions/workspace/ui/side-bar/component-tree/component-tree.tsx
@@ -3,7 +3,8 @@ import memoizeOne from 'memoize-one';
 import { inflateToTree } from './inflate-paths';
 import { TreeNodeContext, TreeNode } from './recursive-tree';
 import { ComponentTreeContext } from './component-tree-context';
-import { /* ScopeView, */ ComponentView, NamespaceView } from './component-nodes';
+import { /* ScopeView, */ NamespaceView } from './component-nodes';
+import { ComponentView } from './component-view';
 import styles from './component-tree.module.scss';
 import { indentStyle } from './indent';
