+++ src/extensions/workspace/ui/side-bar/component-tree/component-view/index.ts
@@ -1,0 +1,1 @@
+export * from './component-view';
